pyproject.toml
[tool.poetry]
name = "network-labs"
version = "0.1.0"
description = "Helper tools for the network-labs repo"
authors = ["Andrew Bates <abates@omeganetserv.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Jinja2 = "^3.1.4"
netaddr = "^1.3.0"
PyYAML = "^6.0.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
